+++ src/stop_command.sh
@@ -1,1 +1,5 @@
-pkill -f "${POD}/odoo/odoo-bin"
+if [[ -f "./conf/odoo.conf" ]]; then
+  pkill -f "${POD}/odoo/odoo-bin"
+else
+  echo "not in a project directory"
+fi
